Move ClickInteractionDetails into interact package
