--- api/src/main/java/com/wolfyscript/utilities/gui/ClickInteractionDetails.java
@@ -1,19 +1,0 @@
-package com.wolfyscript.utilities.gui;
-
-public interface ClickInteractionDetails extends InteractionDetails {
-
-    boolean isShift();
-
-    boolean isSecondary();
-
-    boolean isPrimary();
-
-    int getSlot();
-
-    int getRawSlot();
-
-    int getHotbarButton();
-
-    ClickType getClickType();
-
-}
